requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
scipy>=1.9.0
h5py>=3.7.0
pywavelets>=1.3.0
statsmodels>=0.13.0
Pillow>=9.0.0
xmltodict>=0.13.0
lxml>=4.9.0
openpyxl>=3.0.0
opencv-python>=4.5.0
